refactor: merge activity repository
  * merge ActivityRepositoryFirestore.kt into ActivitiesRepositoryFirestore.kt

  * complete ActivitiesRepository interface with missing functions

  * delete Activity.kt file, which was redundant and uncoherent with the project
