+++ app/src/main/java/com/android/sample/model/activity/ActivitiesRepository.kt
@@ -8,4 +8,8 @@ interface ActivitiesRepository {
   fun getActivities(onSuccess: (List<Activity>) -> Unit, onFailure: (Exception) -> Unit)
 
   fun addActivity(activity: Activity, onSuccess: () -> Unit, onFailure: (Exception) -> Unit)
+
+  fun updateActivity(activity: Activity, onSuccess: () -> Unit, onFailure: (Exception) -> Unit)
+
+  fun deleteActivityById(id: String, onSuccess: () -> Unit, onFailure: (Exception) -> Unit)
 }
